+++ examples/virtual_dpad.rs
@@ -28,7 +28,6 @@ fn spawn_player(mut commands: Commands) {
         .insert(Player)
         .insert_bundle(InputManagerBundle::<Action> {
             // Stores "which actions are currently activated"
-            action_state: ActionState::default(),
             // Map some arbitrary keys into a virtual direction pad that triggers our move action
             input_map: InputMap::new([(
                 VirtualDPad {
@@ -40,6 +39,7 @@ fn spawn_player(mut commands: Commands) {
                 Action::Move,
             )])
             .build(),
+            ..default()
         });
 }