+++ src/lib.rs
@@ -4,6 +4,7 @@
 #![doc = include_str!("../README.md")]
 
 use crate::action_state::ActionState;
+use crate::cooldown::Cooldowns;
 use crate::input_map::InputMap;
 use bevy::ecs::prelude::*;
 use std::marker::PhantomData;
@@ -12,6 +13,7 @@ pub mod action_state;
 pub mod axislike;
 pub mod buttonlike;
 pub mod clashing_inputs;
+pub mod cooldown;
 mod display_impl;
 pub mod errors;
 pub mod input_map;
@@ -31,9 +33,10 @@ pub mod prelude {
     pub use crate::axislike::{DualAxis, MouseWheelAxisType, SingleAxis, VirtualDPad};
     pub use crate::buttonlike::MouseWheelDirection;
     pub use crate::clashing_inputs::ClashStrategy;
+    pub use crate::cooldown::{Cooldown, Cooldowns};
     pub use crate::input_map::InputMap;
     pub use crate::input_mocking::MockInput;
-    pub use crate::user_input::UserInput;
+    pub use crate::user_input::{Modifier, UserInput};
 
     pub use crate::plugin::InputManagerPlugin;
     pub use crate::plugin::ToggleActions;
@@ -132,10 +135,12 @@ impl<A: Actionlike> Default for ActionIter<A> {
 /// Use with [`InputManagerPlugin`](crate::plugin::InputManagerPlugin), providing the same enum type to both.
 #[derive(Bundle)]
 pub struct InputManagerBundle<A: Actionlike> {
-    /// An [ActionState] component
+    /// An [`ActionState`] component
     pub action_state: ActionState<A>,
-    /// An [InputMap] component
+    /// An [`InputMap`] component
     pub input_map: InputMap<A>,
+    /// A [`Cooldowns`] component
+    pub cooldowns: Cooldowns<A>,
 }
 
 // Cannot use derive(Default), as it forces an undesirable bound on our generics
@@ -144,6 +149,7 @@ impl<A: Actionlike> Default for InputManagerBundle<A> {
         Self {
             action_state: ActionState::default(),
             input_map: InputMap::default(),
+            cooldowns: Cooldowns::default(),
         }
     }
 }